lees_edwards: Type conversion

+++ src/core/grid_based_algorithms/lb_walberla_instance.cpp
@@ -94,7 +94,8 @@ bool activate_lb_walberla(std::shared_ptr<LBWalberlaBase> lb_fluid,
     assert(::lattice_switch == ActiveLB::NONE);
     lb_sanity_checks(*lb_fluid, *lb_params, get_time_step());
     auto const &lebc = ::box_geo.lees_edwards_bc();
-    lb_fluid->check_lebc(lebc.shear_direction, lebc.shear_plane_normal);
+    lb_fluid->check_lebc(static_cast<unsigned int>(lebc.shear_direction),
+                         static_cast<unsigned int>(lebc.shear_plane_normal));
   } catch (std::exception const &e) {
     runtimeErrorMsg() << "during waLBerla activation: " << e.what();
     flag_failure = true;
